demo/src/test/java/com/example/demo/service/CustomerServiceImplTest.java
package com.example.demo.service;

import com.example.demo.entity.Customer;
import com.example.demo.repository.CustomerRepository;
import com.example.demo.service.impl.CustomerServiceImpl;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.Arrays;
import java.util.List;
import java.util.Optional;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.Mockito.*;

public class CustomerServiceImplTest {

    @InjectMocks
    private CustomerServiceImpl customerService;

    @Mock
    private CustomerRepository customerRepository;

    @BeforeEach
    public void setup() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    public void testGetAllCustomers() {
        // Arrange
        List<Customer> customers = Arrays.asList(
                new Customer("1", "John Doe", "john@example.com", "123-456-7890"),
                new Customer("2", "Jane Smith", "jane@example.com", "987-654-3210")
        );
        when(customerRepository.findAll()).thenReturn(customers);

        // Act
        List<Customer> result = customerService.getAllCustomers();

        // Assert
        assertEquals(2, result.size());
        verify(customerRepository, times(1)).findAll();
    }

    @Test
    public void testGetCustomerById() {
        // Arrange
        Customer customer = new Customer("1", "John Doe", "john@example.com", "123-456-7890");
        when(customerRepository.findById(Long.valueOf("1"))).thenReturn(Optional.of(customer));

        // Act
        Customer result = customerService.getCustomerById("1");

        // Assert
        assertNotNull(result);
        assertEquals("John Doe", result.getName());
        verify(customerRepository, times(1)).findById(Long.valueOf("1"));
    }

}
